Fix UInt64 SqlParameter crash in ServerHelper.GetServerInfo
ADO.NET doesn't support System.UInt64 as a SQL Server parameter type,
so passing the raw ulong serverId to SqlParameter("@ServerUID", serverId)
threw ArgumentException: The parameter data type of UInt64 is invalid
on every call to ServerHelper.GetServerInfo — which fires every tick-60
for the word-puzzle and jackpot blocks, killing the scheduler loop.

Cast serverId to (long) before wrapping it in SqlParameter, matching
the pattern used everywhere else in the codebase.

+++ Helper/ServerHelper.cs
@@ -13,7 +13,7 @@ namespace DiscordBot.Helper
         public static ServerInfo GetServerInfo(ulong serverId)
         {
             StoredProcedure stored = new StoredProcedure();
-            DataTable dt = stored.Select(Constants.Constants.discordBotConnStr, "GetServerByID", [new SqlParameter("@ServerUID", serverId)]);
+            DataTable dt = stored.Select(Constants.Constants.discordBotConnStr, "GetServerByID", [new SqlParameter("@ServerUID", (long)serverId)]);
             var serverInfo = ServerInfo.PopulateByDataTable(dt);
 
             if (serverInfo != null) return serverInfo;
